+++ src/App.js
@@ -6,19 +6,18 @@ import {
 } from "react-router-dom";
 import "./App.css";
 import Messenger from "./components/Chat/Messenger";
-import LayoutWithVerticalNav from "./components/LayoutWithVerticalNav.js";
 import Navbar from "./components/Navbar/Navbar";
 import AboutUs from "./components/Pages/AboutUs";
 import Faqs from "./components/Pages/Faqs";
 import Feedback from "./components/Pages/Feedback";
 import Vision from "./components/Pages/Vision";
 import Homepage from "./components/Pages/Homepage";
-import Harsh from "./components/Harsh/Harsh";
+import VerticalLayout from "./components/VerticalLayout";
 import SignIn from "./components/Pages/SignIn";
 import SignUp from "./components/Pages/SignUp/SignUp";
-// import Resource from "./components/Resources/Resource.js";
 import { AuthProvider } from "./context/AuthContext";
-import ResourceLinks from "./components/Harsh/ResourceLinks";
+import ResourceLinks from "./components/Resources/ResourceLinks";
+import Resource from "./components/Resources/Resource";
 
 function App() {
   const history = useHistory();
@@ -28,20 +27,18 @@ function App() {
         <Router history={history}>
           <Navbar />
           <Switch>
-            <Route path="/harsh" exact component={Harsh} />
-            <Route path="/harsh/:id" component={ResourceLinks} />
             <Route path="/home" exact component={Homepage} />
             <Route path="/vision" exact component={Vision} />
             <Route path="/aboutus" component={AboutUs} />
             <Route path="/faqs" component={Faqs} />
             <Route path="/signin" component={SignIn} />
             <Route exact path="/register" component={SignUp} />
-            <LayoutWithVerticalNav>
+            <VerticalLayout>
+              <Route path="/resources" exact component={Resource} />
               <Route path="/feedback" component={Feedback} />
+              <Route path="/resources/:id" component={ResourceLinks} />
               <Route path="/chat" component={Messenger} />
-              {/* <Route path="/resource" component={Resource} /> */}
-              {/* <Route path="/:id" component={ResourcePage} />  Change the paths to exact and path for resource to '/resources/:id'  */}
-            </LayoutWithVerticalNav>
+            </VerticalLayout>
           </Switch>
         </Router>
       </AuthProvider>
