--- src/components/VerticalNav/VerticalNav.css
@@ -1,40 +1,0 @@
-.vertical-nav {
-  width: 8rem;
-  height: calc(100% - 1rem);
-  margin: auto 0.6rem;
-  border-radius: 13px;
-  background-color: rgba(128, 185, 232, 1);
-  padding: 2rem 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-}
-.profile-pic-container {
-  width: 3.7rem;
-  height: 3.7rem;
-  border-radius: 50%;
-  background-color: white;
-  overflow: hidden;
-  margin-bottom: 1rem;
-}
-.profile-pic {
-  height: 100%;
-}
-.vertical-nav-btns-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.vertical-nav-icon {
-  height: 27px;
-  margin-bottom: 2.1rem;
-}
-.log-out-icon {
-  height: 27px;
-  cursor: pointer;
-}
-
-.chat-icon{
-  /* height: 35px; */
-}
